Reverted to non-momentum-based movement
Still able to use wasd  and arrow keys to move.
Note: If you're moving and cast the speed spell, you'll move at the same slow speed until you press a movement key again. Also true if you've casted the speed spell and are moving quickly. If you don't release the key, you'll keep moving quickly in that direction even if the spell wore off.
Explanation: The function on_key() only runs when a key is pressed. This means that it will only run once when a key is pressed and, unlike world_system.step(), will not refresh. This means it won't be able to see any changes to player speed until a new key is pressed. This is circumvented with a flag system and having step() handle updating movement. A brief attempt was made, but due to time constraints, could not be realized.
Also note: holding a movement key then tapping another movement key will cause the player to stop moving despite a movement key still being held.