Added ability to destroy enemy on attack and others
Expected Behavior:
Pressing space while face the monster will destroy it. This allows the player to navigate the maze more freely for the time being
Restrictions: enemy can't be directly on top of player, they should be ~50px within center of player.
Enemy has to be in front of the direction the player is facing. Tho for now just having x position greater / lesser than player is good enough.

Other changes
Moved spawnable_tiles vector outside of loop, this was an attempt to implement a respawn feature, which is too buggy for showcase.
Replaced MAX_TURTLES and MAX_FISH to MAX_DRONES and MAX_SPIKES. Respawn feature was going to use this